refactor: reorganize pipeline layers and move jepx raw uploader
