+++ src/pipeline/scraper/main.py
@@ -3,7 +3,7 @@ import logging
 
 from core.storage_client import RustFSClient
 from pipeline.jepx.common import resolve_target_at
-from pipeline.scraper.jepx_to_rustfs import scrape_jepx_to_rustfs
+from pipeline.raw.jepx_to_rustfs import scrape_jepx_to_rustfs
 from pipeline.scraper.module.jepx import JEPXSpotSummaryScraper
 
 logging.basicConfig(
